+++ transforms-python/src/myproject/datasets/procedure_occurrence.py
@@ -30,7 +30,7 @@ def compute(ctx, omop_eav_dict):
         .withColumn('procedure_datetime',  F.to_timestamp(F.col('procedure_datetime'))) \
         .withColumn('procedure_type_concept_id', df['procedure_type_concept_id'].cast(T.IntegerType())) \
         .withColumn('quantity', df['quantity'].cast(T.IntegerType()))\
-        .withColumn('data_partner_id', df['data_partner_id'].cast(T.LongType())) \
+        .withColumn('data_partner_id', df['data_partner_id'].cast(T.IntegerType())) \
 
 
     df = df.select([
